tests/requirements.txt
pytest~=8.3.5
requests~=2.32.3
python-dotenv~=1.1.0
confluent-kafka~=2.10.0
minio~=7.2.15
pymongo~=4.12.1